[Common] fix the Network Interface tests for Windows
There was a problem with the tests where the reported NIC name was not
found abd didn't pass validation.
Now the search for the NIC for windows will return all values and will
test that at least one is valid.
In addition, will test the logic when `ifaddr` package is not installed.

Signed-off-by: DjLegolas <djlegolas@protonmail.com>